+++ src/main/java/nordmods/uselessreptile/common/item/VariantChangingOrbItem.java
@@ -1,7 +1,9 @@
 package nordmods.uselessreptile.common.item;
 
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.chat.Component;
+import net.minecraft.util.CommonColors;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.LivingEntity;
@@ -42,7 +44,7 @@ public class VariantChangingOrbItem extends Item {
 
     @Override
     public void appendHoverText(@NonNull ItemStack itemStack, @NonNull TooltipContext tooltipContext, @NonNull TooltipDisplay tooltipDisplay, @NonNull Consumer<Component> consumer, @NonNull TooltipFlag tooltipFlag) {
-        consumer.accept(Component.translatable("tooltip.uselessreptile.variant_changing_orb"));
+        consumer.accept(Component.translatable("tooltip.uselessreptile.variant_changing_orb").withStyle(ChatFormatting.GRAY));
         super.appendHoverText(itemStack, tooltipContext, tooltipDisplay, consumer, tooltipFlag);
     }
 }